Rename `kernel::shape` to `operations`
The near-collision with `kernel::shape` has always been unfortunate, so
renaming one of them is a good idea. I think `operations` is a good name
for the module, as that's what it's focused on: Translating the
operations from the `fj` crate into something the kernel understands.
