+++ fj-app/src/kernel/operations/circle.rs
