+++ fj-app/src/kernel/operations/difference_2d.rs
