+++ fj-app/src/kernel/operations/group.rs
