+++ fj-app/src/kernel/operations/mod.rs
