+++ fj-app/src/kernel/operations/sketch.rs
